pyproject.toml
[project]
name = "experiment-knowornot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "google-generativeai>=0.8.5",
    "knowornot",
    "playwright>=1.52.0",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
knowornot = { path = "../KnowOrNot" }
